refactor(core): group theme support
