+++ src/core/theme/legacySyntaxScopes.test.ts
